requirements.lock.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
authlib==1.6.1
bandit==1.8.6
boolean-py==5.0
cachecontrol==0.14.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dparse==0.6.4
fastapi==0.116.1
filelock==3.16.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jinja2==3.1.6
joblib==1.5.1
license-expression==30.4.4
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
msgpack==1.1.1
mypy==1.17.0
mypy-extensions==1.1.0
nltk==3.9.1
packageurl-python==0.17.1
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
pip==25.1.1
pip-api==0.0.34
pip-audit==2.9.0
pip-requirements-parser==32.0.1
platformdirs==4.3.8
psutil==6.1.1
py-serializable==2.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
ruff==0.12.5
safety==3.6.0
safety-schemas==0.0.14
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.41
starlette==0.47.2
stevedore==5.4.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
